+++ src/wfg/native_ui/ui/system/UISystemContainer.java
@@ -63,7 +63,7 @@ public class UISystemContainer {
 
 
     public final void invalidateCache() { cached = null;}
-    public final List<BaseSystem> getCustomSystems() { return Collections.unmodifiableList(cached);}
+    public final List<BaseSystem> getCustomSystems() { return Collections.unmodifiableList(customSystems); }
     public final List<BaseSystem> getAll() {
         if (cached == null) {
             cached = new ArrayList<>();